+++ src/roq/okx/static_data.cpp
@@ -379,7 +379,7 @@ void StaticData::operator()(Trace<json::Instruments> const &event) {
           .symbol = symbol,
           .description = {},
           .security_type = map(item.inst_type),
-          .external_security_id = {},
+          .external_security_id = utils::safe_cast(item.inst_id_code),
           .cfi_code = {},
           .base_currency = base_currency,
           .quote_currency = quote_currency,