+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 1.1.7 &ndash; 2026-08-08
+
 ### Fixed
 
 * Only disconnect the public WS connections when the `service_type` is WS (#615)